+++ src/components/dropdown/dropdown.js
@@ -1,5 +1,5 @@
 import * as debug from '../../utils/debug';
-import { deprecateMethod, warnAboutDeprecation } from '../../utils/deprecated';
+import { deprecateMethod } from '../../utils/deprecated';
 import { utils } from '../../utils/utils';
 import { DOM } from '../../utils/dom';
 import { Environment as env } from '../../utils/environment';
@@ -3039,14 +3039,10 @@ Dropdown.prototype = {
 
   /**
    * Public API for setting the `<select>`'s value.
-   * This method is slated to be removed in a future v4.10.0 or v5.0.0.
    * @private
-   * @deprecated as of v4.4.0. It's preferrable to set the input value and call `updated()` on the component API.
    * @param {string} code - The value to match and set on the value element.
    */
   setCode(code) {
-    warnAboutDeprecation(this.setCode, this.element.val);
-
     const self = this;
     const doSetting = function () {
       self.element.val(code);